requirements.txt
Django==3.1
djangorestframework==3.11.1
psycopg2-binary
pandas
xlrd==1.2.0
requests
tqdm
gunicorn==19.9.0
django-cors-headers
fiona
pyshp
matplotlib
tqdm